Several performance improvements
- Change CcValueInj to have a list of containers.
- Change return value of 'inj-creation' methods from a list of injs to
one injection.
- KappaParser now accept expressions as init values.
- Reimplementation of NullaryOperations (to handle [time]).
- Fix bugs in Plot.h.
- Reimplementation of Warnings.
- Bug fixing with tokens.

+++ src/util/params.h
@@ -4,7 +4,7 @@
 
 //#define DEBUG true //in make command
 
-typedef float FL_TYPE;
+typedef double FL_TYPE;
 typedef long INT_TYPE;
 typedef unsigned long UINT_TYPE;
 #define FRMT_FL "f"
